Count rate metric for transaction/transaction trees streams

+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/config/WorkflowConfig.scala
@@ -67,7 +67,7 @@ object WorkflowConfig {
     case class Objectives(
         maxDelaySeconds: Option[Long],
         minConsumptionSpeed: Option[Double],
-        minItemRate: Option[Long],
+        minItemRate: Option[Double],
     )
   }
 }
